// 요청 작동 구현을 위해 firestore의 users에 true,false값 들어가게 설정(기본 값은 false) // null -> no-null 어쩌구 오류 수정

+++ app/src/main/java/com/example/cho6/guru2_android12/friendList.kt
@@ -240,10 +240,10 @@ class FriendListViewModel : ViewModel() {
                 .addOnSuccessListener { result ->
                     for (document in result) {
                         if (document.id != user.uid) {    //로그인한 본인 데이터가 아닌 다른사람들
-                            val yourLatitude = document.data["latitude"] as Double
+                            val yourLatitude = document.data["latitude"] as Double 
                             val yourLongitude = document.data["longitude"] as Double
                             val distance: Double =
-                                getDistance(myLatitude, myLongitude, yourLatitude, yourLongitude)
+                                getDistance(myLatitude, myLongitude, yourLatitude.toDouble(), yourLongitude)
                             if (distance <= 300) {
                                 Log.d("test", "" + document.data["username"])
                                 data.add(
